Fixed #19; added a button to delete all tasks, and delete all completed tasks

+++ app/src/main/java/com/wcsm/minhastarefas/MainActivity.kt
@@ -142,8 +142,10 @@ class MainActivity : AppCompatActivity() {
             if(taskList.isNotEmpty()) {
                 val taskDAO = TaskDAO(this)
                 taskList.forEach {
-                    if(!taskDAO.delete(it.id)) {
-                        deletingErrors++
+                    if(it.completed == 0) {
+                        if(!taskDAO.delete(it.id)) {
+                            deletingErrors++
+                        }
                     }
                 }
                 if(deletingErrors > 0) {
